% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{cluster_tab}
\alias{cluster_tab}
\title{split a dtm into two clusters with reinert algorithm}
\usage{
cluster_tab(dtm, cc_test = 0.3, tsj = 3, ...)
}
\description{
split a dtm into two clusters with reinert algorithm
}
