% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_uc}
\alias{compute_uc}
\title{Merges uces into uc according to minimum uc size}
\usage{
compute_uc(dtm, min_uc_size = 10)
}
\arguments{
\item{dtm}{dtm of uces}

\item{min_uc_size}{minimum number of forms by uc}
}
\value{
the original dtm with a new `rainette_uc_id` docvar.
}
\description{
`rainette_uc_index` docvar
}
