% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree.R
\name{cutree.rainette}
\alias{cutree.rainette}
\title{Cut a rainette result tree into groups of documents}
\usage{
cutree.rainette(hres, k = NULL, h = NULL)
}
\arguments{
\item{hres}{the `rainette` result object to be cut}

\item{k}{the desired number of groups}

\item{h}{unsupported}
}
\description{
Cut a rainette result tree into groups of documents
}
