% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree.R
\name{cutree}
\alias{cutree}
\title{Cut a tree into groups}
\usage{
cutree(tree, ...)
}
\arguments{
\item{tree}{the hclust tree object to be cut}

\item{...}{arguments passed to other methods}
}
\description{
Cut a tree into groups
}
\details{
If `tree` is of class `rainette`, invokes `cutree.rainette`. Otherwise, just run `stats::cutree`.
}
