% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree.R
\name{cutree_rainette}
\alias{cutree_rainette}
\title{Cut a rainette result tree into groups of documents}
\usage{
cutree_rainette(hres, k = NULL, h = NULL, ...)
}
\arguments{
\item{hres}{the \code{rainette} result object to be cut}

\item{k}{the desired number of groups}

\item{h}{unsupported}

\item{...}{arguments passed to other methods}
}
\value{
A vector with group membership.
}
\description{
Cut a rainette result tree into groups of documents
}
