% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{docs_order_by_ca}
\alias{docs_order_by_ca}
\title{return documents indices ordered by CA first axis coordinates}
\usage{
docs_order_by_ca(dtm, ...)
}
\arguments{
\item{dtm}{dmt on which to compute the CA and order documents}

\item{...}{arguments passed to \code{quanteda::textmodel_ca}}
}
\value{
ordered list of document indices
}
\description{
return documents indices ordered by CA first axis coordinates
}
\details{
Internal function, not to be used directly
}
