% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{features_selection}
\alias{features_selection}
\title{Remove features from dtm of each group base don cc_test and tsj}
\usage{
features_selection(tab, indices1, indices2, cc_test, tsj)
}
\arguments{
\item{tab}{global dtm}

\item{indices1}{indices of documents of group 1}

\item{indices2}{indices of documents of group 2}

\item{cc_test}{cc_test value}

\item{tsj}{tsj value}
}
\value{
a list of two character vectors : `cols1` is the name of features to 
keep in group 1, `cols2` the name of features to keep in group 2
}
\description{
Remove features from dtm of each group base don cc_test and tsj
}
