% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_corpus_iramuteq}
\alias{import_corpus_iramuteq}
\title{Import a corpus in Iramuteq format}
\usage{
import_corpus_iramuteq(f, id_var = NULL, thematics = c("remove", "split"), ...)
}
\arguments{
\item{f}{a file name or a connection}

\item{id_var}{name of metadata variable to be used as documents id}

\item{thematics}{if "remove", thematics lines are removed. If "split", texts as splitted at each thematic, and metadata duplicated accordingly}

\item{...}{arguments passed to \code{\link[base:file]{base::file()}} if \code{f} is a file name.}
}
\value{
A quanteda corpus object. Note that metadata variables in docvars are all imported as characters.
}
\description{
Import a corpus in Iramuteq format
}
\details{
A description of the Iramuteq corpus format can be found here : \url{http://www.iramuteq.org/documentation/html/2-2-2-les-regles-de-formatages}
}
