% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_segments}
\alias{merge_segments}
\title{Merges uces into uc according to minimum uc size}
\usage{
merge_segments(dtm, min_segment_size = 10, doc_id = NULL)
}
\arguments{
\item{dtm}{dtm of uces, with a \code{rainette_uce_id} docvar}

\item{min_segment_size}{minimum number of forms by uc}

\item{doc_id}{character name of a dtm docvar which identifies source documents.}
}
\value{
the original dtm with a new \code{rainette_uc_id} docvar.
}
\description{
\code{rainette_uc_index} docvar
}
\details{
If \code{min_segment_size == 0}, different uc ids are added to the dtm docvars
(ie no uce are merged together). If \code{min_segment_size > 0} then \code{doc_id} must be provided
unless the corpus comes from \code{split_segments}, in this case
\code{segment_source} is used by default.
}
