% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{order_docs}
\alias{order_docs}
\title{return documents indices ordered by CA first axis coordinates}
\usage{
order_docs(m)
}
\arguments{
\item{m}{dtm on which to compute the CA and order documents, converted to an integer matrix.}
}
\value{
ordered list of document indices
}
\description{
return documents indices ordered by CA first axis coordinates
}
\details{
Internal function, not to be used directly
}
