% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree.R
\name{rainette2_complete_groups}
\alias{rainette2_complete_groups}
\title{Complete groups membership with k-means}
\usage{
rainette2_complete_groups(dfm, groups, ...)
}
\arguments{
\item{dfm}{dfm object used for `rainette2`` clustering}

\item{groups}{group membership computed by \code{cutree} on \code{rainette2} result.}

\item{...}{other arguments passed to \code{kmeans}}
}
\description{
Starting with groups membership computed from a \code{rainette2} clustering,
every document not assigned to a cluster is reassigned using a k-means
clustering initialized with the centers of defined groups.
}
\seealso{
\code{\link[=cutree_rainette2]{cutree_rainette2()}}
}
