% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette2_explor.R
\name{rainette2_explor}
\alias{rainette2_explor}
\title{Shiny gadget for rainette2 clustering exploration}
\usage{
rainette2_explor(res, dtm = NULL, corpus_src = NULL)
}
\arguments{
\item{res}{result object of a \code{rainette2} clustering}

\item{dtm}{the dfm object used to compute the clustering}

\item{corpus_src}{the quanteda corpus object used to compute the dtm}
}
\value{
No return value, called for side effects.
}
\description{
Shiny gadget for rainette2 clustering exploration
}
\seealso{
\code{\link[=rainette2_plot]{rainette2_plot()}}
}
