% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rainette2_plot}
\alias{rainette2_plot}
\title{Generate a clustering description plot from a rainette2 result}
\usage{
rainette2_plot(res, dtm, k = NULL, criterion = c("chi2", "n"),
  type = c("bar", "cloud"), n_terms = 15, free_scales = FALSE,
  measure = c("chi2", "lr"), show_negative = TRUE, text_size = NULL)
}
\arguments{
\item{res}{result object of a `rainette2` clustering}

\item{dtm}{the dfm object used to compute the clustering}

\item{k}{number of groups. If NULL, use the biggest number possible}

\item{criterion}{criterion to use to choose the best partition. `chi2` means 
the partition with the maximum sum of chi2, `n` the partition with the 
maximum size.}

\item{type}{type of term plots : barplot or wordcloud}

\item{n_terms}{number of terms to display in keyness plots}

\item{free_scales}{if TRUE, all the keyness plots will have the same scale}

\item{measure}{statistics to compute}

\item{show_negative}{if TRUE, show negative keyness features}

\item{text_size}{font size for barplots, max word size for wordclouds}
}
\description{
Generate a clustering description plot from a rainette2 result
}
\seealso{
`quanteda::textstat_keyness`, `rainette2_explor`
}
