% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rainette_plot}
\alias{rainette_plot}
\title{Generate a clustering description plot from a rainette result}
\usage{
rainette_plot(res, dtm, k = NULL, type = c("bar", "wordcloud"),
  n_terms = 15, free_x = FALSE, measure = c("chi2", "lr"),
  font_size = NULL)
}
\arguments{
\item{res}{result object of a `rainette` clustering}

\item{dtm}{the dfm object used to compute the clustering}

\item{k}{number of groups. If NULL, use the biggest number possible.}

\item{n_terms}{number of terms to display in keyness plots}

\item{free_x}{if TRUE, all the keyness plots will have the same x axis}

\item{measure}{statistics to compute}

\item{font_size}{font size for terms plots}
}
\description{
Generate a clustering description plot from a rainette result
}
\examples{
\dontrun{
library(quanteda)
corpus <- data_corpus_inaugural
corpus <- head(corpus, n = 10)
corpus <- split_segments(corpus)
dtm <- dfm(corpus, remove = stopwords("en"), tolower = TRUE, remove_punct = TRUE)
dtm <- dfm_trim(dtm, min_termfreq = 3)
res <- rainette(dtm, k = 3)
rainette_plot(dtm, res)
}

}
\seealso{
`quanteda::textstat_keyness`, `rainette_explor`
}
