% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{split_tab_by_chisq}
\alias{split_tab_by_chisq}
\title{Split a dtm in two according to the chi-square value between the two aggregated tables}
\usage{
split_tab_by_chisq(tab, indices)
}
\arguments{
\item{tab}{dtm to be split}

\item{indices}{documents indices ordered by CA first axis coordinate}
}
\value{
a list with the index of the document where the split generates a maximum chi-square,
and the corresponding max chi-square value
}
\description{
Split a dtm in two according to the chi-square value between the two aggregated tables
}
\details{
Internal function, not to be used directly
}
