% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{switch_docs_by_chisq}
\alias{switch_docs_by_chisq}
\title{Switch documents between two groups to maximize chi-square value}
\usage{
switch_docs_by_chisq(tab, indices, max_index, max_chisq)
}
\arguments{
\item{tab}{original dtm}

\item{indices}{documents indices orderes by first CA axis coordinates}

\item{max_index}{document index where the split is maximum, computed with \code{\link[=split_tab_by_chisq]{split_tab_by_chisq()}}}

\item{max_chisq}{maximum chi-square value computed by \code{\link[=split_tab_by_chisq]{split_tab_by_chisq()}}}
}
\value{
a list of two vectors \code{indices1} and \code{indices2}, which contain
the documents indices of each group after documents switching, and a \code{chisq} value,
the new corresponding chi-square value after switching
}
\description{
Switch documents between two groups to maximize chi-square value
}
\details{
Internal function, not to be used directly
}
