% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{cluster_tab}
\alias{cluster_tab}
\title{Split a dtm into two clusters with reinert algorithm}
\usage{
cluster_tab(dtm, cc_test = 0.3, tsj = 3)
}
\arguments{
\item{dtm}{to be split, passed by \code{rainette}}

\item{cc_test}{maximum contingency coefficient value for the
feature to be kept in both groups.}

\item{tsj}{minimum feature frequency in the dtm}
}
\value{
An object of class \code{hclust} and \code{rainette}
}
\description{
Split a dtm into two clusters with reinert algorithm
}
\details{
Internal function, not to be used directly
}
