% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree.R
\name{cutree}
\alias{cutree}
\title{Cut a tree into groups}
\usage{
cutree(tree, ...)
}
\arguments{
\item{tree}{the hclust tree object to be cut}

\item{...}{arguments passed to other methods}
}
\value{
A vector with group membership.
}
\description{
Cut a tree into groups
}
\details{
If \code{tree} is of class \code{rainette}, invokes \code{\link[=cutree_rainette]{cutree_rainette()}}. Otherwise, just run \code{\link[stats:cutree]{stats::cutree()}}.
}
