% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainette.R
\name{rainette}
\alias{rainette}
\title{Performs a corpus clustering based on the Reinert method}
\usage{
rainette(dtm, k = 10, min_uc_size = 10, min_members = 5,
  cc_test = 0.3, tsj = 3, verbose = FALSE, ...)
}
\arguments{
\item{dtm}{dtm of documents to cluster}

\item{k}{number of clusters to compute}

\item{min_uc_size}{minimum number of forms by document}

\item{min_members}{don't try to split groups with fewer members}

\item{cc_test}{contingency coefficient value for feature selection}

\item{tsj}{minimum frequency value for feature selection}

\item{verbose}{if TRUE, add intermediate computing elements to results object}

\item{...}{parameters passed to `quanteda::textmodel_ca`}
}
\value{
TODO : compléter
}
\description{
Performs a corpus clustering based on the Reinert method
}
\details{
TODO : compléter
}
\examples{
\dontrun{
library(quanteda)
corpus <- data_corpus_inaugural
corpus <- head(corpus, n = 10)
corpus <- split_segments(corpus)
dtm <- dfm(corpus, remove = stopwords("en"), tolower = TRUE, remove_punct = TRUE)
dtm <- dfm_wordstem(dtm, language = "english")
dtm <- dfm_trim(dtm, min_termfreq = 3)
res <- rainette(dtm, k = 3)
}

}
\references{
Reinert M, Une méthode de classification descendante hiérarchique : application à l'analyse lexicale par contexte, Cahiers de l'analyse des données, Volume 8, Numéro 2, 1983. [http://www.numdam.org/item/?id=CAD_1983__8_2_187_0](http://www.numdam.org/item/?id=CAD_1983__8_2_187_0)

Reinert M., Alceste une méthodologie d'analyse des données textuelles et une application: Aurelia De Gerard De Nerval, Bulletin de Méthodologie Sociologique, Volume 26, Numéro 1, 1990. [https://doi.org/10.1177/075910639002600103](https://doi.org/10.1177/075910639002600103)
}
