% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_segments.R
\name{split_segments}
\alias{split_segments}
\title{Split a character string into segments, taking into account punctuation where possible}
\usage{
split_segments(text, segment_size = 40, segment_size_window = 15)
}
\arguments{
\item{text}{character string}

\item{segment_size}{segment size in words}

\item{segment_size_window}{window around segment size to look for best splitting point}
}
\value{
A tibble with the computed segments
}
\description{
Split a character string into segments, taking into account punctuation where possible
}
